feat: Adds customizable detection area for barcode scanner (#17)
* feat: Adds customizable detection area for barcode scanner

Implements functionality to customize the barcode detection area.

Allows specifying the width and height of the detection area via options passed to the `scan` method.  This can be useful for optimizing scanning for different barcode sizes and aspect ratios.

Adds examples to the documentation, and ensures a minimum size for the detection area for usability.

* chore: Renames constants

Renames the constants for the detection area size to include "DEFAULT"
to clarify their purpose and improve code readability.

* Updates version to 1.7.0

Adds customizable detection area for the barcode scanner.